fix OOM errors with newer LRU
with large item support, we switched from tracking the number of chunks each
sub-LRU had to the amount of bytes. This was undercounting the bytes by
skipping the header, which could matter enough for the smaller slab classes to
allow LRU_HOT to envelop COLD, preventing evictions from happening.

This wasn't made too obvious by the fact that causing this requires an
(unusually?) even slab page distribution.