NOEXP_LRU can now act as a "transient" LRU
Been resisting this for a long time due to how HOT operates. However it's
probably a good way to relieve some pressure off the LRU crawler and do fewer
cross-LRU juggles.

The "NOEXP" LRU can still function by simply setting the transient_ttl to 0
(handy!). This is a test so the variables haven't been renamed yet, and
there's a perhaps temporary command for live-modification of the transient
threshold.