LRU maint test fix for second mystery
If an LRU is too large, ACTIVE processing was elided on purpose. However it's
very easy for even moderate write pressure to pass a large number of ACTIVE
items into COLD. This combined with not removing items from COLD immediately
(fixed in another commit) could easily cause very active items to be evicted
along with inactive ones.

With the speedup of the processing thread (with some low-fruit still), it
should be fine to quickly flip items in place.

This opens up a problem under extreme write pressure which overwhelms the
background thread, or extreme re-scanning of large numbers of items.

It might still be valuable to do pressure release or direct reclaim pulling if
COLD gets too large. Performance tests should be done to find the threshold
and experiment.