Avoid segfault if idle_timeout value is missing.
If memcached is called without a numeric argument for the idle_timeout option, so:
  -o idle_timeout
and not:
  -o idle_timeout=60
it segfaults inside stdlib/strtol_l.c:: ____strtol_l_internal from stdlib.h::atoi from the IDLE_TIMEOUT case.
I am using glibc-2.12 on RHEL6.  I think atoi(NULL) is undefined behaviour.
This change guards against this for the the idle_timeout option, in the same manner as the cases that use the subopts_value, printing that the numeric argument is missing for the idle_timeout option.