default to unix sockets for tests
Tests have randomly failed to start since the beginning of time, due largely
to port assignment. A random local port is tested, then passed onto the daemon
to try. If that port (on 0.0.0.0) becomes used in the meantime, it fails.

Now, tests default to using unix sockets unique to the pid of the test parent.
Some tests still run via the network and have been changed to 127.0.0.1, which
should collide less with normal internet connections. Some tests require that
due to some timing issues that unix sockets seem to create, and others expect
a network in a few cases.

Ran the tests with `PARALLEL=9 make test` on my machine for two hours and they
didn't fail once.